Move gl rendering code to render_vertices
